+++ CHANGELOG.md
@@ -9,6 +9,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
  ### Added
  ### Changed 
  ### Fixed 
+* Fix AQL handling for "FOLDER CONTAINS VERSION CONTAINS COMPOSITION" [#1459](https://github.com/ehrbase/ehrbase/pull/1459)
 
 ## [2.12.0]
  ### Added