+++ CHANGELOG.md
@@ -5,13 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.13.0]
+ ### Added
+ ### Changed 
+ ### Fixed 
+* Fix AQL handling for "FOLDER CONTAINS VERSION CONTAINS COMPOSITION" [#1459](https://github.com/ehrbase/ehrbase/pull/1459)
+
 ## [2.12.0]
  ### Added
  ### Changed 
 * Added check for duplicate version IDs during Folder creation as well as check for Folder uid and if-match header id during update ([#1410](https://github.com/ehrbase/ehrbase/pull/1410))
 * Remove EHR from AQL if not needed for the query ([#1448](https://github.com/ehrbase/ehrbase/pull/1448))
+* Update flyway to V11 to Gracefully exit in case installed flyway DB is newer than supported ([#1450](https://github.com/ehrbase/ehrbase/pull/1450))
  ### Fixed 
-* Allow Template overwrite with renamed property from `system.allow-template-overwrite` to `ehrbase.template..allow-overwrite` ([#1440](https://github.com/ehrbase/ehrbase/pull/1440))
+* Allow Template overwrite with renamed property from `system.allow-template-overwrite` to `ehrbase.template.allow-overwrite` ([#1440](https://github.com/ehrbase/ehrbase/pull/1440))
 
 ## [2.11.0]
  ### Added
@@ -135,3 +142,4 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 [2.10.0]: https://github.com/ehrbase/ehrbase/compare/v2.9.0...v2.10.0
 [2.11.0]: https://github.com/ehrbase/ehrbase/compare/v2.10.0...v2.11.0
 [2.12.0]: https://github.com/ehrbase/ehrbase/compare/v2.11.0...v2.12.0
+[2.13.0]: https://github.com/ehrbase/ehrbase/compare/v2.12.0...v2.13.0